# HPCC-13922 Speed up --pq option for the regression suite