@ghalliday 2015-07-16
@richardkchapman @AttilaVamos please review.  (I doubt I have got it completely correct)

Some sample timings, running locally with --pq 20

before:
roxie: 537s
hthor: 548s

after:
roxie: 325s
hthor: 202s

Note, this is an improvement, but the code should really be refactored and further improved.
- there are 3 places that are calling thread.start_new_thread.  They should all call a common function.
- there shouldn't really be any timing waits - the code should be using a semaphore instead.
- I suspect more could be done to improve the latency
- if compiles were split from execution it should be able to speed up the sequential execution as well.