@HPCCSmoketest 2015-07-16
Automated Smoketest Build: success
HPCC Start: OK
HPCC Stop: OK