@richardkchapman 2015-07-17
@AttilaVamos please review