@AttilaVamos 2015-07-17
Checked, tested, @ghalliday thanks for this improvement. 
It is good to merge.